+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v.1.0.5
+
+### Under the hood
+
+- add FireboltColumn class to handle string type
+
 ## v.1.0.4
 
 ### Features